backend/requirements.txt
flask
gunicorn
numpy
requests
psycopg2-binary